Refactors geometry traits and improves data structures
This commit refactors the geometry traits by extracting the `OrderedEq` trait into its own module for better organization and reusability.

It also modifies data structures to use `const fn` where applicable, potentially improving performance and enabling compile-time evaluation in certain contexts.

Additionally, it adds missing words to the spellcheck dictionary.
