+++ cspell.json
@@ -25,9 +25,11 @@
         "cpus",
         "EATCS",
         "Edelsbrunner",
+        "finitecheck",
         "fuga",
         "Getchell",
         "getrandom",
+        "hashcoordinate",
         "hashset",
         "hypertetrahedron",
         "itoa",
@@ -38,6 +40,7 @@
         "memchr",
         "nalgebra",
         "nocapture",
+        "orderedeq",
         "powi",
         "Preparata",
         "profraw",
